backend/requirements.txt
# Backend Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiofiles==23.2.1

# Database
sqlalchemy==2.0.23
alembic==1.13.0

# Vector Database
chromadb==0.4.18
numpy<2.0.0  # ChromaDB 需要 NumPy 1.x
opencv-python-headless==4.8.1.78  # 兼容 NumPy 1.x

# AI/ML
openai==1.3.7
sentence-transformers==2.2.2
imagehash==4.3.1
pillow==10.1.0

# Task Scheduling
apscheduler==3.10.4

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
pydantic==2.5.2
pydantic-settings==2.1.0
